+++ apps/docs-v2/pages/products/index.tsx
@@ -1,7 +1,7 @@
 import {Box, Column, Typography} from '@comfortdelgro/react-compass'
 import AppFooter from 'components/layouts/AppFooter'
-import ProductService from 'components/products/product-services'
-import ProductsBackground from 'components/products/products-background'
+import ProductService from 'components/products/ProductServices'
+import ProductsBackground from 'components/products/ProductsBackground'
 import Head from 'next/head'
 
 const Products = () => {
